feat: Use original commit message on one commit updates
If a template update only applies one commit to the
target repository, the original commit message and
body can be applied without modifications
instead of the summary with all changes for bulk edits.

This change introduces a detection if the template
update only contains one commit and uses the
original message accordingly.

This will allow maintainers to update repositories
in smaller increments, such that the commit history
may partially mirror the commit history of the template.